backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
django-filter>=23.0
django-cors-headers>=4.3
